+++ classes/output/core_renderer_toolbox.php
@@ -44,7 +44,13 @@ trait core_renderer_toolbox {
     public function render_page() {
         global $CFG;
 
+        $toolbox = \theme_foundation\toolbox::get_instance();
         $mustache = $this->page->theme->layouts[$this->page->pagelayout]['mustache'];
+        $trio = (!empty($toolbox->get_setting('trio')));
+        if ($trio && ($mustache == 'columns2')) {
+            $mustache = 'columns3';
+        }
+
         $data = new \stdClass();
         $data->output = $this;
         $bodyclasses = array();
@@ -94,9 +100,24 @@ trait core_renderer_toolbox {
                 $data->sidepreblocks = $preblockshtml;
                 $data->haspreblocks = $haspreblocks;
             }
+
+            if (in_array('side-post', $this->page->theme->layouts[$this->page->pagelayout]['regions'])) {
+                $postblockshtml = $this->blocks('side-post');
+                $haspostblocks = ((strpos($postblockshtml, 'data-block=') !== false) or ($this->page->user_is_editing()));
+
+                $data->sidepreblocks = $postblockshtml;
+                $data->haspostblocks = $haspostblocks;
+            }
+
+            if ($trio) {
+                if (!empty($data->haspreblocks)) {
+                    $data->hasblocks = true;
+                } else if (!empty($data->haspostblocks)) {
+                    $data->hasblocks = true;
+                }
+            }
         }
 
-        $toolbox = \theme_foundation\toolbox::get_instance();
         $modules = $toolbox->get_modules();
         foreach ($modules as $module) {
             if (method_exists ($module, 'export_for_template')) {