+++ Changes.md
@@ -5,6 +5,7 @@ Version 3.10.1.1 - TBR
 1. Fix header opacity settings not resetting the theme cache.
 2. Added swatch custom colour detection.
 3. Added core H5P alter styles support - MDL-69087.
+4. Add three column layout capability from Classic.
 
 Version 3.10.1.0 - 04/03/21
 ---------------------------